feat: separation of @deroll/app into @deroll/app and @deroll/core
