+++ .changeset/green-students-taste.md
@@ -1,0 +1,11 @@
+---
+"eslint-config-deroll": minor
+"minimal": minor
+"wallet": minor
+"@deroll/router": minor
+"@deroll/wallet": minor
+"@deroll/core": minor
+"@deroll/app": minor
+---
+
+separation of @deroll/app into @deroll/app and @deroll/core
