+++ examples/wallet/package.json
@@ -6,22 +6,21 @@
         "build": "tsup",
         "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
         "dev": "tsup --watch",
-        "lint": "eslint",
-        "test": "vitest"
+        "lint": "eslint"
     },
     "dependencies": {
         "@deroll/app": "workspace:*",
         "@deroll/router": "workspace:*",
         "@deroll/wallet": "workspace:*",
-        "viem": "^2.9.9"
+        "viem": "^2.9.19"
     },
     "devDependencies": {
         "@deroll/tsconfig": "workspace:*",
-        "@types/node": "^20.12.5",
+        "@types/node": "^20.12.7",
         "eslint": "^8.57.0",
         "eslint-config-deroll": "workspace:*",
         "tsup": "^8.0.2",
-        "typescript": "^5.4.3",
-        "vitest": "^1.4.0"
+        "typescript": "^5.4.5",
+        "vitest": "^1.5.0"
     }
 }
