+++ package.json
@@ -17,7 +17,7 @@
         "eslint-config-deroll": "workspace:*",
         "prettier": "^3.2.5",
         "turbo": "^1.13.2",
-        "typedoc": "^0.25.12"
+        "typedoc": "^0.25.13"
     },
-    "packageManager": "pnpm@8.15.5"
+    "packageManager": "pnpm@9.0.1"
 }
