--- packages/app/.gitignore
@@ -1,1 +1,0 @@
-/src/schema.d.ts
