+++ packages/core/schema.ts
@@ -7,8 +7,9 @@ Node API defined at https://openapi-ts.pages.dev/node/. The goal is to use the
 viem types Hex and Address instead of simple strings for some schema properties.
 */
 
-const inputFile = "rollup.yaml";
-const outputFile = "src/schema.d.ts";
+const inputFile =
+    "https://raw.githubusercontent.com/cartesi/openapi-interfaces/v0.7.2/rollup.yaml";
+const outputFile = "src/schema.ts";
 
 // import types from viem in generated code
 const inject = "import { Address, Hex } from 'viem';\n";
