+++ packages/core/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "@deroll/tsconfig/base.json",
+    "compilerOptions": {
+        "noUncheckedIndexedAccess": true
+    },
+    "include": ["."],
+    "exclude": ["dist", "node_modules"]
+}
