+++ packages/eslint-config-deroll/package.json
@@ -5,7 +5,7 @@
     "license": "MIT",
     "private": true,
     "dependencies": {
-        "@typescript-eslint/eslint-plugin": "^7.5.0",
+        "@typescript-eslint/eslint-plugin": "^7.7.0",
         "eslint-config-prettier": "^9.1.0",
         "eslint-config-turbo": "^1.13.2",
         "eslint-plugin-import": "^2"
