+++ packages/router/package.json
@@ -19,21 +19,21 @@
         "prepack": "run-s build"
     },
     "dependencies": {
-        "@deroll/app": "workspace:*",
-        "path-to-regexp": "^6",
-        "viem": "^2.9.9"
+        "@deroll/core": "workspace:*",
+        "path-to-regexp": "^6.2.2",
+        "viem": "^2.9.19"
     },
     "devDependencies": {
         "@deroll/tsconfig": "workspace:*",
-        "@types/node": "^20.12.5",
+        "@types/node": "^20.12.7",
         "eslint": "^8.57.0",
         "eslint-config-deroll": "workspace:*",
         "npm-run-all": "^4",
         "openapi-typescript": "^6.7.5",
         "ts-node": "^10.9.2",
         "tsup": "^8.0.2",
-        "typescript": "^5.4.3",
-        "vitest": "^1.4.0",
+        "typescript": "^5.4.5",
+        "vitest": "^1.5.0",
         "vitest-mock-extended": "^1.3.1"
     },
     "publishConfig": {
