+++ packages/wallet/package.json
@@ -20,23 +20,23 @@
         "test": "vitest"
     },
     "dependencies": {
-        "@deroll/app": "workspace:*",
-        "viem": "^2.9.9"
+        "@deroll/core": "workspace:*",
+        "viem": "^2.9.19"
     },
     "devDependencies": {
-        "@cartesi/rollups": "^1.2",
+        "@cartesi/rollups": "^1.3.1",
         "@deroll/tsconfig": "workspace:*",
         "@sunodo/wagmi-plugin-hardhat-deploy": "^0.3",
-        "@vitest/coverage-istanbul": "^1.4.0",
-        "@vitest/ui": "^1.4.0",
+        "@vitest/coverage-istanbul": "^1.5.0",
+        "@vitest/ui": "^1.5.0",
         "@wagmi/cli": "^2.1.4",
         "abitype": "^1.0.2",
         "eslint": "^8.57.0",
         "eslint-config-deroll": "workspace:*",
         "npm-run-all": "^4",
         "tsup": "^8.0.2",
-        "typescript": "^5.4.3",
-        "vitest": "^1.4.0"
+        "typescript": "^5.4.5",
+        "vitest": "^1.5.0"
     },
     "publishConfig": {
         "access": "public"
