+++ packages/wallet/src/wallet.ts
@@ -1,4 +1,4 @@
-import { AdvanceRequestHandler, Voucher } from "@deroll/app";
+import { AdvanceRequestHandler, Voucher } from "@deroll/core";
 import {
     Address,
     Hex,
@@ -35,8 +35,8 @@ type DeepReadonly<T> = Readonly<{
     [K in keyof T]: T[K] extends number | string | symbol
         ? Readonly<T[K]>
         : T[K] extends Array<infer A>
-        ? Readonly<Array<DeepReadonly<A>>>
-        : DeepReadonly<T[K]>;
+          ? Readonly<Array<DeepReadonly<A>>>
+          : DeepReadonly<T[K]>;
 }>;
 
 const createEmptyWallet = (): Wallet => ({
